Apply ACCELERATION_TEMPORAL to ACCELERATION_RAMPING.
This is a first draft.

The simple idea is, calculate ramping as before on the fast axis,
but apply the result to all four axes. Timings for a slow axis
can be found by a simple comparison with the fast axis.

Try the algorithm by #defining ACCELERATION_RAMPING. The
ACCELERATION_TEMPORAL flag is now obsolete.

As always with first drafts, there are a number of obvious
optimisations not applied to keep the diff small and readable.